# Conflicts:
#	calculator/src/main/java/com/market/banica/calculator/service/CalculatorServiceImpl.java
#	calculator/src/main/java/com/market/banica/calculator/service/TestData.java

+++ calculator/src/main/java/com/market/banica/calculator/service/grpc/AuroraClientSideService.java
@@ -52,7 +52,7 @@ public class AuroraClientSideService {
         }
     }
 
-    public ItemOrderBookResponse getIngredient(String productName, String clientId, Integer quantity) {
+    public ItemOrderBookResponse getIngredient(String productName, String clientId, Long quantity) {
         LOGGER.debug("Inside getIngredient method with parameter product name - {} and client id - {}"
                 , productName, clientId);
